fix loop fusion load bug

+++ loop_fusion/src/pose_graph.cpp
@@ -416,7 +416,7 @@ int PoseGraph::detectLoop(KeyFrame* keyframe, int frame_index)
         int min_index = -1;
         for (unsigned int i = 0; i < ret.size(); i++)
         {
-            if (min_index == -1 || (ret[i].Id < min_index && ret[i].Score > 0.015))
+            if (min_index == -1 || (static_cast<int>(ret[i].Id) < min_index && ret[i].Score > 0.015))
                 min_index = ret[i].Id;
         }
         return min_index;
